Fixing bug of inserting 0
Since we are using multiplicative hashing, it is better to start inserting keys
in the data structure from 1. The random key generator also generate keys in
the range [1..N], so it would happen that a key '0' would never be generated
and accesses to key 'N' would always fail (key would not exist).