Fixing overflow bug in zipfian distribution
The parameter n in the zeta function had the same type as the one specified for
the distribution object.
The constructor calls the zeta function with n == (max-min+1).
If the range specified was the whole domain of the given integer type, using
the same type for the n parameter in zeta would cause an overflow (max-0+1).
We fix that by using the largest possible integer type for n.