Setting default option values inline
We want libpibench to be linked by third-party applications, rather than just
the PiBench binary.

Before the handling of default values for options was being handled in the
PiBench binary (main.cpp). It is better to have the default values inlined in
the declaration, since this allows both PiBench binary and other applications
to get the default values directly.

Consequently, we removed all handling of default options from main.cpp.