Replace reserve() by resize()
While std::vector::reserve() pre-allocates memory, it does not initialize it,
which leads to page faults during execution when values are assigned for the
first time.

An alternative is to call resize(MAX_SIZE) followed by resize(0) to achieve the
same effect, but forcing the page faults before the measuremenets start.