Replacing discrete generator by uniform for ops
It turns out that the discrete distribution is more expensive than the regular
uniform one.
Alternatively we populate a fixed-size array of 256 operations based on the
weights of each operation when constructing the operation_generator and then
generate a random position in this array during runtime for picking an op.